+++ src/lib.rs
@@ -53,7 +53,7 @@ pub fn example() {
         .text("Test Label")
         .place(&mut window);
 
-    y += label.rect.get().height as isize + 10;
+    y += label.rect.get().height as i32 + 10;
 
     let text_box = TextBox::new()
         .position(x, y)
@@ -64,7 +64,7 @@ pub fn example() {
         .place(&mut window);
 
     let button = Button::new()
-        .position(x + text_box.rect.get().width as isize + 10, y)
+        .position(x + text_box.rect.get().width as i32 + 10, y)
         .size(48, text_box.rect.get().height)
         .text("Update")
         .on_click(move |_button: &Button, _point: Point| {
@@ -72,7 +72,7 @@ pub fn example() {
         })
         .place(&mut window);
 
-    y += button.rect.get().height as isize + 10;
+    y += button.rect.get().height as i32 + 10;
 
     let progress_label = Label::new()
         .text("Progress: 0%")
@@ -80,19 +80,19 @@ pub fn example() {
         .size(400, 16)
         .place(&mut window);
 
-    y += progress_label.rect.get().height as isize + 10;
+    y += progress_label.rect.get().height as i32 + 10;
 
     let progress_bar = ProgressBar::new()
         .position(x, y)
         .size(400, 16)
         .on_click(move |progress_bar: &ProgressBar, point: Point| {
-            let progress = point.x * 100 / progress_bar.rect.get().width as isize;
+            let progress = point.x * 100 / progress_bar.rect.get().width as i32;
             progress_label.text.set(format!("Progress: {}%", progress));
             progress_bar.value.set(progress);
         })
         .place(&mut window);
 
-    y += progress_bar.rect.get().height as isize + 10;
+    y += progress_bar.rect.get().height as i32 + 10;
 
     let multi_line_label = Label::new()
         .text("Multi-Line Text")
@@ -100,7 +100,7 @@ pub fn example() {
         .size(400, 16)
         .place(&mut window);
 
-    y += multi_line_label.rect.get().height as isize + 10;
+    y += multi_line_label.rect.get().height as i32 + 10;
 
     TextBox::new()
         .position(x, y)
