+++ src/place.rs
@@ -1,4 +1,4 @@
 pub trait Place {
-    fn position(self, x: isize, y: isize) -> Self;
-    fn size(self, width: usize, height: usize) -> Self;
+    fn position(self, x: i32, y: i32) -> Self;
+    fn size(self, width: u32, height: u32) -> Self;
 }
